requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
ConfigArgParse==1.7
Django==4.2.3
django-dbml==0.6.0
django-sslserver==0.22
djangorestframework==3.14.0
et-xmlfile==1.1.0
gevent==23.9.1
geventhttpclient==2.0.10
greenlet==3.0.0rc3
idna==3.4
itsdangerous==2.1.2
locust==2.16.1
MarkupSafe==2.1.3
msgpack==1.0.5
mysqlclient==2.2.1
openpyxl==3.1.2
Pillow==10.0.0
psutil==5.9.5
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.4
zope.event==5.0
zope.interface==6.0
